Nfs linking service now ignores files without -r mod

+++ features.in/03unroot/live/files/usr/local/bin/start-profiling.sh
@@ -7,6 +7,7 @@ DEFAULT_FILES=(
 	".gitconfig"
 )
 
+# Создаём дефолтные файлы, если их нет, выставляем папке .ssh необходимые права
 for file in "${DEFAULT_FILES[@]}"; do
 	if [ ! -e "$BACKUP_DIR/$file" ]; then
 		if [ "$file" == ".ssh/" ]; then
@@ -18,16 +19,19 @@ for file in "${DEFAULT_FILES[@]}"; do
 	fi
 done
 
+# заменяем файлы из домашней директории на ссылки на соответствующие файлы из $BACKUP_DIR
 find "$BACKUP_DIR" -maxdepth 1 | while read -r item; do
-	file=$(basename $item)
-	if [ -d "$item" ]; then
-		rm -rf "$HOME/$file"
-		ln -s "$item" "$HOME"
-	fi
-	if [ -f "$item" ]; then
-		rm -rf "$HOME/$file"
-		ln -s "$item" "$HOME/$file"
-		echo "dir $item $file"
+	if [ -r $item ]; then
+		file=$(basename $item)
+		if [ -d "$item" ]; then
+			rm -rf "$HOME/$file"
+			ln -s "$item" "$HOME"
+		fi
+		if [ -f "$item" ]; then
+			rm -rf "$HOME/$file"
+			ln -s "$item" "$HOME/$file"
+			echo "dir $item $file"
+		fi
 	fi
 done
 exit 0
